+++ src/solve/commonfunctions.cpp
@@ -23,10 +23,10 @@
 // graph library
 #include "graph.hpp"
 
-graph::Edge findBottleneck(const graph::Euclidean& euclidean, const std::vector<unsigned int>& tour, const bool isCycle) {
-  unsigned int bottleneckEdgeEnd = 0;
+graph::Edge findBottleneck(const graph::Euclidean& euclidean, const std::vector<size_t>& tour, const bool isCycle) {
+  size_t bottleneckEdgeEnd = 0;
   double bottleneckWeight        = euclidean.weight(tour[0], tour[1]);
-  for (unsigned int i = 1; i < euclidean.numberOfNodes() - 1; ++i) {
+  for (size_t i = 1; i < euclidean.numberOfNodes() - 1; ++i) {
     if (euclidean.weight(tour[i], tour[i + 1]) > bottleneckWeight) {
       bottleneckEdgeEnd = i;
       bottleneckWeight  = euclidean.weight(tour[i], tour[i + 1]);
